Patch set includes primarily cosmetic clean-up of the group member. The
substanative change is the unification of the on-disk and in-memory group member
state.

* 'group-member-v2' of github.com:noahdesu/v:
  kafka: deinline exceptional path
  kafka: use signed ints for group/member accounting
  kafka: simply some member tests
  kafka: move largish member impl out of header
  kafka: add member state serde test
  kafka: iobuf uses need to be in seastar context
  kafka: unify in-memory and on-disk member state
  bytes: add bytes-to-iobuf conversion helpers
  reflect: support std::chrono::millisecond serde
  kafka: use absl flat hash set for protocol candidates
  kafka: rename group_member::vote to be meaningful
  kafka: rename the group_member::metadata method
  kafka: remove for_each_protocol member accessor
  kafka: simplifyy group member protocol matching

+++ src/v/kafka/groups/group.h
@@ -412,7 +412,7 @@ private:
     kafka::generation_id _generation;
     protocol_support _supported_protocols;
     member_map _members;
-    size_t _num_members_joining;
+    int _num_members_joining;
     absl::flat_hash_set<kafka::member_id> _pending_members;
     std::optional<kafka::protocol_type> _protocol_type;
     std::optional<kafka::protocol_name> _protocol;
